C3M0_A: "Background" zombie optimization
- Added check to Nazi class for a custom sector property: 'user_background'.  Make enemies use a generic wandering state and run less code when in sectors with the property set.

If user_background is set to True or 1 for a sector, any Nazi-descendant enemies in the sector will skip most of the new game logic and will fall back to the previously unused "Idle" state - so they'll wander around and play their active sound, and that's it, as long as they are in a 'user_background'-tagged sector.

This lets us keep the hordes of civilian zombies active outside of the city gates with around a third of the think time that they had when running all code as was previously normal.