Attempt to fix MG Turret soldier bugs
Add +DONTCORPSE to the MGTurret, so that the game will not go to the Crash state upon dying, since the MGTurret is already on the ground.
Add ControllableBase, which is a base class for actors which can be "controlled" - they check for "controller" actors within range, and replace themselves with other actors.
Remove Wehrmacht variant of CTurret_Afrika, and replace CTurret_Afrika with a MG turret that can be "controlled" by both Wehrmacht and Afrika Korps guards.
Make MechaNaziStanding actors inherit ControllableBase