Make Nemesis death/resurrection controllable with ACS
Remove unused GroundedEnergyBolt and NemesisZyklonCloud actors. GroundedEnergyBolt was just there for testing, and NemesisZyklonCloud was unused.
Add NemesisControl thinker, which stores bools to control whether the Nemesis can die, or remain kneeling.
Add pointer to NemesisControl, and initialize it when he is spawned.
Make Nemesis use the NemesisControl to decide whether to die, or "sleep".