Rewrite and fix ClampAngle
The new version of ClampAngle uses Actor.deltaangle to tell the angle difference between the min and max angles, and so it is more robust than a plain angle comparison. Hopefully, this won't need any further tweaks, and maybe I'll try to add it as a native GZDoom engine feature.