Surrendering weak troops
This implements my idea written in #322 about surrendering weak regular units (only mechanics and scientists, no soldiers, of course). When only one of them remains in the room and is cornered by the player, he will drop his weapon and surrender. Done in quite a crude manner, improvements are very welcome.
Is there any way to remove the growing red blood spot on the actor's death? Or, better yet, to change the liquid colour to yellow :-D