Inventory Bar Adjustments
- Only useable items will now show up in the inventory bar (so there are now only a max of 13 items to scroll through)
- All other miscellaneous items (CompassItems and PuzzleItems) are drawn in the top right of the screen in columns of 5 - under the key bar in fullscreen, and all the way in the corner on standard status bar
- Added the ability for CompassItems to remove 'alternate' items from theinventory when they are given - e.g., the two halves of the Egyptian Artifact are removed when the whole shield is given
- Added a 'give money' cheat shortcut to give gold, and added gold to the 'give all' handling so that you get max money with the cheat also

The item icon handling uses the same code as the current inventory selection drawing, so the amount numbers are also drawn as applicable.

This handling works for all items (so if you 'give all', everthing draws properly), but most levels won't need anywhere near the full number of items to draw, ever.