C3M0_A: Various fixes
- Changed one RE actor (small baluster - there are 194 in the map) to inherit from SimpleActor for performance reasons
- Tweaked breakable statue code to make it more reliable if something was underneath the statue when it fell.
- Updated ACS script to remove skull key from player inventory when used
- Minor geometry change in manor to eliminate portal glitch wierdness (single-line-thick wall in double stairwells)
- Teaked the shape and angle of the gem placement when it is used

Issue that I've added to the tracker: There is still an opportunity for the player to get stuck at (9800, 2800) if they went straight into the excavation without going through the steps to get C4.  You can jump down here and not be able to return or advance.

A side note: I had originally intended the statue Gem to be used for a non-required secret  in the mansion (gaining the ZyklonResistance vial, which is intentionally unlabeled and gives no indication of what it does - but gives you 75% resistance to Zyklon damage).  This implementation feels out of place to me; the door almost feels like it needs a more Hexen-like 'gear in the wall' approach.