Fix max ammo after removing RyanToken
After becoming James Ryan, your maximum ammo counts would be reset because you lost your backpack if you had it. This is an attempt to fix that.
Also, add HQTraining, which is preliminary, but it is meant for the training course at Allied HQ, and it will hold the player's inventory like RyanToken.