C3M2: Re-Added 1/3 and 2/3 stars for the "3 target" objective in the map
@Talon1024 I noticed something strange and I hope you can help with it.

If I pukename BATTLEFIELD_OBJECTIVES 1-3 in a random order, everything works fine.

If I play the map, first destroy the communication and second destroy the zyklon c tanks, the 2/3 star is not appearing, I have absolutely no idea why this is not working but puking the scripts is. Maybe you have an idea.