Update fontchars.csv with all characters used in language.csv
I wrote a (Python) script which scans language.csv, and shows me which characters are currently being used, as well as their uppercase and lowercase forms. I used this new script alongside a zsh script which scans fontchars.csv, and each font, for missing characters in each font.
On another note, I think this is a faster, and possibly more reliable solution than @AFADoomer's ingame language.csv checks. His script said that the "typewriter" and "MavenProSmall" fonts were missing the "ß" character, when they actually are not.