UMG43 altfire rework
This is what I proposed in https://github.com/Realm667/WolfenDoom/issues/303#issuecomment-766692301. Only needs proper sprites now (including a high-res four-barrel gunflash).
The alternative attack mode can kill normal enemies in one pull of the trigger (4*random(11,15) is larger than health of most Wehrmacht/SS enemies), it will probably stun larger enemies in pain state (especially super-soldiers), and it is easy to conserve ammo with it. I also made it very accurate, so the four-bullet pattern can be easily seen. To compensate for this I lowered the DPS by around 2x, but I guess more serious measures need to be taken (though UMG43 is the best weapon in the game, and is quite rare...) So, @Tormentor667 and @Ozymandias81, please share your thoughts on balancing it.
@Ozymandias81, could you also please convert the sounds to ogg?