Add a large ammo cache to C3M0_A
See https://github.com/Realm667/WolfenDoom/issues/383#issuecomment-783844378.
The lift is now opened with a switch behind Hitler in Totengraeber's room, and it leads to a small basement with weapons and ammo (via static portals). They should not be a performance issue here, since sight in the directions of large-sized areas is blocked by one-sided lines.
I made the basement obligatory to visit (there are two tag 383 enemies in it).