Changes to dragging feature
It is left off, but now it would be much easier to switch it on in an addon.
The ACS needs to be recompiled, but there is also a problem which I commented on: you can sprint (and actually receive a 2x speed bonus) when the crouch is toggled. Hold-to-crouch button prevents sprinting, as it should.