Prevent player from buying a Kar98k before C2M5
The player is given a Kar98k at the start of C2M5, so it doesn't make much sense to be able to buy one before C2M5.
Add two new conversations for Lt. Rojas to the INTERMAP DIALOGUE lump: One which is activated before C2M5, one which is activated after C2M5, and make use of them in the NPCDialogue script.