Custom option menu handling for language selection
- Added code to display the percentage complete for each language as selected, if it is below 100%

There's no additional message being displayed to explain the number, but that would be trivial to add, just need the LANGUAGE entries for that line in each language.

This is...  Kind of hacky, but is the only way I know of to change only one specific menu control while maintaining full compatibility with the built-in menus.