Move BoA_PlayerLoop to BoAPlayer
This looping script was impossible to switch off, as ENTER scripts can not be TERMINATEd.
I moved all its logic to BoAPlayer's Tick(). Had to provide ugly script wrappers, since there is no interface to call ACS-defined functions from ZScript, but this way at least we can decide if we want it to run or not.
I tested this commit, everything works properly and no behaviour is changed (including other playerpawns), but you can see it for yourself ;-)