ZScriptfied sprint and footstep generation
- These are built as inventory items so that they can be given/taken easily, and for easy portability of code to other mods if wanted.

If you notice any differences in behavior, please let me know; I ported the ACS directly in, then did some optimizations.

Note that this also eliminates the need for terrain-specific footstep splashes to be defined, instead using the (undocumented and unused in GZDoom) LeftStepSounds and RightStepSounds properties of the terrain definitions to set up the sounds.