Rework sprinting again to achieve the intended effect
Rename AffectSideMove to AnyDirection
Remove some bAffectSideMove checks

If the player is not holding down the sprint key, the player should be moving at normal speed in the desired direction.
If the player IS holding down the sprint key, the player should be moving twice as fast in the desired direction.
If bAnyDirection is false, the forward/backward AND lateral speed should only be multiplied if the player is holding down the forward move key.