Slightly rework revolving doors, and fix the one on C1M5
The revolving doors now treat the original line's arg1 for lines with rotation (and rotation override) specials as the torque multiplier.
Also, add specials to the edges of the revolving door so that, if it is rotated in such a way that it blocks the "Hotel de la Mort" entrance, the player can press use on the end of it, and it should move out of the player's way to allow them to enter.

This should fix #593, although I'm not sure if this is the best solution.