C2M6_B: Allow progression via alternate route
If you jumped over the railing to get into the water with the mines (~6400, 3200) and had not yet collected the red key, you were soft locked.  This change makes the underwater switch also open the standard path back to the rest of the map, since it was previously only openable from the "intended" side.

Should address #732.