Slightly rework InventoryHolder and tokens
Add armorIcon and armorHeld fields to InventoryHolder so that the icon is restored
Don't make RyanToken or HQTrainingCourse call Super.TryPickup; instead, call AttachToOwner directly
Add Inventory.NeverRespawn flags so that DisguiseToken and HQTrainingCourse are never respawned - they are tokens

This should fix #806. The reason players left the training course with 0% health is because the owner was assigned by AttachToOwner, which was called by TryPickup AFTER their inventory had been held.