Make the SSToilet2 sneakable
I used a "spy camera" to find out that the SSToilet2 was moving around frantically because the guard patrolling the hallway decided to shoot him when he opened the bathroom door. The guard had the FRIENDLY flag set, but the SSToilet2 did not, and I think it has something to do with that.

Fixes #1266