Particle ash and cinders
The actors were removed, and all instances of them converted to A_SpawnParticleEx. The method SpawnCinder can be moved to a class common to all effect actors (and TexMan lookups can be made to be done only once).
More to come later (if anyone wants to help by editing these and copy-pasting into files of /effects, you're welcome :-) )
@Ozymandias81, could you please comment on the cinders (to see them you can spawn OvenHot or kill Guards with the flamer)? The new function uses a different convention for size, and I did not adjust it 1:1.