Partial fix for #1244
Removed the lockup, but now it is possible to launch Ryan into air using a flamethrower, and he will stay there until you order him to come with you. Why does not gravity work on him? Apparently the two SetOrigin calls in the file are not the cause of this..?
@Talon1024, could you please help with this issue?