Attempt to fix #1115 for good
If the player raised the first lift from the bottom, and then ran off of it, the lift would get stuck at the top if the player flipped the switch to call the lift before the lift was done moving because the variable was being changed before the lift was done moving.

I also added two switches to call the second lift.