Complete rewrite of line breaking and colouring
I had to do this because the new code only checks whether or not to break the line when there's whitespace, but the old code checked whether or not to break the line at every character.

I did a bit of testing, and there's no issues as far as I can tell.