Various map fixes
C2M1 - Now the handrails render flush with the wall.
C3M0_A - Reinstated the texture build_b49 on line 1122. Was causing an issue above the windows outside the mansion. 
C3M2 - Now the handrails render flush with the wall and floor of sector 909 alignment fix.
C3M3_A - The backside of the deer skull alignment. Line 10161 for example.