C3M1_A Fixes/Changes/Improvements
1. Polyobject sectors 2160 and 2202 are now sliding sectors. They were swing sectors before.
Reason for change is a) When opened, part of the door would render into the ceiling of slope/plane sector 2172 and 2161 and
b) Would pass through the red light at sectors 2203 and 2171. Red light moved to cell.
A torture, makes it hard for the inmate to sleep ...
2. I omitted to fix the crossbeams on the 4 guard towers east. Now fixed.
3. At sector 5038 look south-west. A rework of the large crate. New things added here - T4658 and T4659.
4. At last the gas drums render on the truck T2642 - and not below.
5. New sector 3674.
6. Before, when at sector 6487 and looking out the windows, you would see the same thing.
Caused by both lines tagged 191. Fixed.
7. Also, when at sector 6487, and other windows in this area, there was a gap in the ceiling
Reworked to fix.
7. A number of floor and ceiling offset fixes.
8. Some trolley and drum repositioning.