Fix issue #1471
Previous fix would have normally worked, but in this case that Thing had a health of -1, which made it an instant kill regardless.
This was the case even in BoA 3.1.
While at it I added Custom Invisible Bridges for things with no hitbox. Many of these were tagged for removal so I tagged  thes CIB's as well.
And flagged the barracks windows block monster line of sight.